bsd.debugdata.mk: define _INSTALL_UNSTRIPPED and add a first stub of post-install-strip-debugdata
Define _INSTALL_UNSTRIPPED internally. This avoid to pass strip options to
several tools. Also document that behaviour for the on-line documentation
(TODO: it's worth to also mention that in the INSTALL_UNSTRIPPED documentation).

Define a first stub of post-install-strip-debugdata that will be probably used
to summons readelf, objcopy, etc. to strip the debug data and generate the
`.debug' files.
XXX: this will be subject to changes and probably several
XXX: targets will be needed to accomplish that.