bsd.debugdata.mk: do not append to _ALL_INSTALL_TARGETS
This is probably just ignored and so not needed.
bsd.debugdata.mk is included later in bsd.pkg.mk. Other `.mk' files in mk/ that
append targets to post-install does not do that.

XXX: pax.mk seems the only exception to that. We should check if it's just not
XXX: needed and in that case get of rid of it.