bsd.debugdata.mk: implement PKG_DEBUGLEVEL functionality
PKG_DEBUGLEVEL can be "small", "default" or "detailed" (if it's not the
following ones the default is "default"). After a quick inspection of the
documentation of gcc and Clang it seems that only gcc have the concept of a
level for the `-g' option.

TODO: transform-gcc wrapper should be instructed regarding -g1 and -g3 options!