check-debugdata.mk: add a sketch version to check object files and .debug files
Like most other check this is triggered only for PKG_DEVELOPERs and if
PKG_DEBUGDATA is yes.
Several part of the code was stolen from check-stripped.mk and check-shlibs.mk
(but it's much more similar to the former for the spirit).

TODO: ATM non-${OBJ_FMT} files aren't properly handled, this is a bug and should
TODO: be fixed ASAP!

XXX: (note to myself) check if the code can be simplified. Unfortunately putting
XXX: the logic in separate script (shell or AWK) is probably not easy because in
XXX: this context using DELAYED_*_MSG is particularly handy.
XXX: If this can be easily obtained or complicate the logic without any real
XXX: gains it will be probably better to provide two separate targets.
XXX: One to just check programs/libraries. Another one to check for the
XXX: `.debug*' sections in `.debug' files.

XXX: Probably `_check-debugdata' and `*DEBUGDATA*' is abused in target and
XXX: variable names. A better terminology to separate debugdata (`.debug' files)
XXX: and program/libraries files is a possible improvement.