bsd.check-vars.mk: add CHECK_DEBUGDATA_SUPPORTED
Used by check-debugdata.mk.

TODO: ATM it's only supported for NetBSD, so setting it to "yes" is a lie!
TODO: (the only tool needed is "objdump" and for non-ELF ${OBJ_FMT} we should
TODO: just worry if they use `.gnu_debuglink'). If that's true, the PLATFORM is
TODO: a real CHECK_DEBUGDATA_SUPPORTED.