check-debugdata.mk: check only file formats that are recognized by objdump
Previously all files where checked and so it lead to problems for everything
that wasn't recognised by objdump. Check the objdump output in order to filter
and ignore all files that aren't recognised.

XXX: still needs to test that `into the wild'!

XXX: A similar logic can be probably used in bsd.debugdata.mk in order to get
XXX: rid of DEBUGDATA per-package variable!