plist.mk: continue to SUBPACKAGES-ification and add TODO
SUBPACKAGES-ify other part of the code (still incomplete).

Add various TODO (TODOleot) that for the moment just address documentation.
A closer look also to `The pkgsrc guide' will be needed when the functionality
will be implemented.

Mark PLIST_TYPE dynamic not compatible with SUBPACKAGES.

XXX: Currently PLIST_TYPE dynamic is not used in pkgsrc...
XXX: Probably we can delete it?