package.mk: Fix completely wrong code and TODOleot affecting `su-package-install'.
In a TODOleot comment I was suggesting that +CONTENTS file could be replaced
with ${_CONTENTS_FILE} defined in metadata.mk. This suggestion was completely
wrong and also the existing code for the SUBPACKAGES was wrong.
All of that was due the confusion between ${_PKG_DBDIR} and ${PKG_DB_TMPDIR}.
The former is usually just `/var/db/pkg' the latter is just a temporary
directory in WRKDIR (`${WRKDIR}/.pkgdb').